+++ tests/core_tests/chaingen.cpp
@@ -45,6 +45,7 @@
 #include "cryptonote_basic/cryptonote_basic_impl.h"
 #include "cryptonote_basic/cryptonote_format_utils.h"
 #include "cryptonote_core/miner.h"
+#include "loki_economy.h"
 
 #include "chaingen.h"
 #include "device/device.hpp"
@@ -533,7 +534,7 @@ cryptonote::transaction loki_chain_generator::create_loki_name_system_tx(crypton
   uint64_t new_height           = get_block_height(top().block) + 1;
   uint8_t new_hf_version        = get_hf_version_at(new_height);
   if (burn == LNS_AUTO_BURN)
-    burn = lns::burn_requirement_in_atomic_loki(new_hf_version, type);
+    burn = lns::burn_needed(new_hf_version, type);
 
   crypto::hash prev_txid = crypto::null_hash;
   if (lns::mapping_record mapping = lns_db_.get_mapping(type, name))
